Use original package structure (#15)
