+++ docs/guide/src/docs/asciidoc/introduction.adoc
@@ -11,19 +11,6 @@ format and it's history, refer to the respective
 https://en.wikipedia.org/wiki/RIS_(file_format)[Wikipedia article].
 
 JRis was initially developed by Gianluca Colaianni (fastluca) and
-is available on github: https://github.com/fastluca/JRis.
-
-This https://github.com/ursjoss/JRis[fork] KRis is currently used
-to develop a variant of the initial version with some differences:
-
-- the code was converted to kotlin and refactored to avoid using
-  reflection
-- the API was developed to easily use the project as a library
-  both for software written in Kotlin and/or Java.
-- The library should be published to bintray and/or maven central,
-  so client software can access the library as a maven or gradle
-  dependency.
-
-It is yet to be decided whether to reintegrate the current changes
-into the original project or whether to keep the fork as KRis
-separately.
+is available on github: https://github.com/fastluca/JRis. It was
+ported to kotlin by Urs Joss (ursjoss) in a fork and is currently
+in the process of being merged again with the original project.
