+++ subprojects/jris/src/main/kotlin/com/gmail/gcolaianni5/jris/RisRecord.kt
